OCM-20819 | CNV-56734 | feat: prevent vCPU overcommit
AWS is legally responsible for having limited and accurate vCPU
allocation for Microsoft License-Included customers on AWS EC2. Having a
misallocation of vCPU would result in a breach of agreement between AWS
and Microsoft for License-Included Windows software. This change deploys
VAP on HCPs that prevents the overcommit of vCPUs for windows VMs.